Refactor asset folder structure
