+++ webapp/src/assets/icons/theta/CapitalAccumulation.tsx
