+++ webapp/src/assets/icons/theta/PrincipalProtection.tsx
