+++ webapp/src/assets/icons/theta/Theta.tsx
