+++ webapp/src/assets/icons/theta/Volatility.tsx
